use directory node for addin file creation

+++ RevitJournal/Revit/Addin/AddinManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Autodesk.RevitAddIns;
+using DataSource.Model.FileSystem;
 using RevitAction;
 
 namespace RevitJournal.Revit.Addin
@@ -11,15 +12,16 @@ namespace RevitJournal.Revit.Addin
     {
         public const string VendorId = "RascerDev";
 
-        public static void CreateAppManifest(string outputPath, ITaskAppInfo info)
+        public static void CreateAppManifest(DirectoryNode directory, ITaskAppInfo info)
         {
             if (info is null) { throw new ArgumentNullException(nameof(info)); }
+            if (directory is null) { throw new ArgumentNullException(nameof(directory)); }
 
             if (string.IsNullOrEmpty(info.VendorId))
             {
                 info.VendorId = VendorId;
             }
-            var manifestPath = GetManifestPath(outputPath);
+            var manifestPath = GetManifestPath(directory);
             if (File.Exists(manifestPath) == false)
             {
                 File.Delete(manifestPath);
@@ -33,11 +35,12 @@ namespace RevitJournal.Revit.Addin
             manifest.SaveAs(manifestPath);
         }
 
-        public static void CreateManifest(string outputPath, ITaskInfo info)
+        public static void CreateManifest(DirectoryNode directory, ITaskInfo info)
         {
             if (info is null) { throw new ArgumentNullException(nameof(info)); }
+            if (directory is null) { throw new ArgumentNullException(nameof(directory)); }
 
-            var manifestPath = GetManifestPath(outputPath);
+            var manifestPath = GetManifestPath(directory);
             var manifest = AddInManifestUtility.GetRevitAddInManifest(manifestPath);
             var commands = manifest.AddInCommands;
             if (HasCommand(commands, info.Id) == false)
@@ -55,9 +58,9 @@ namespace RevitJournal.Revit.Addin
             }
         }
 
-        private static string GetManifestPath(string outputPath)
+        private static string GetManifestPath(DirectoryNode directory)
         {
-            return Path.Combine(outputPath, "RevitJournalAddins.addin");
+            return Path.Combine(directory.FullPath, "RevitJournalAddins.addin");
         }
 
         private static bool HasCommand(IEnumerable<RevitAddInCommand> commands, Guid guid)
